Extract a AppFlow component to contain the flow
